backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-dotenv==1.0.0
supabase==1.0.3
requests==2.31.0
numpy==1.24.3
scikit-learn==1.2.2
